requirements.txt
black==22.8.0
Django==4.1.1
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
flake8==5.0.4
Markdown==3.4.1
phonenumbers==8.12.56
psycopg2-binary==2.9.3
python-dotenv==0.21.0
Pillow==9.2.0
